% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_line.R
\name{ChartLine}
\alias{ChartLine}
\title{ChartLine}
\format{
\code{\link{R6Class}} object, super class \code{\link{ChartSize}}.
}
\description{
ChartLine

ChartLine
}
\section{Super classes}{
\code{\link[tercenApi:Base]{tercenApi::Base}} -> \code{\link[tercenApi:Chart]{tercenApi::Chart}} -> \code{\link[tercenApi:ChartSize]{tercenApi::ChartSize}} -> \code{ChartLine}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pointSize}}{of type int inherited from super class \code{\link{ChartSize}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Chart}}.}

\item{\code{properties}}{object of class \code{\link{Properties}} inherited from super class \code{\link{Chart}}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{pointSize}}{of type int inherited from super class \code{\link{ChartSize}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Chart}}.}

\item{\code{properties}}{object of class \code{\link{Properties}} inherited from super class \code{\link{Chart}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ChartLine$new()}}
\item \href{#method-init}{\code{ChartLine$init()}}
\item \href{#method-initJson}{\code{ChartLine$initJson()}}
\item \href{#method-toTson}{\code{ChartLine$toTson()}}
\item \href{#method-print}{\code{ChartLine$print()}}
\item \href{#method-clone}{\code{ChartLine$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartLine$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartLine$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartLine$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartLine$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartLine$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChartLine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
