% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{Column}
\alias{Column}
\title{Column}
\format{
\code{\link{R6Class}} object, super class \code{\link{ColumnSchema}}.
}
\description{
Column

Column
}
\section{Super classes}{
\code{\link[tercen:Base]{tercen::Base}} -> \code{\link[tercen:IdObject]{tercen::IdObject}} -> \code{\link[tercen:ColumnSchema]{tercen::ColumnSchema}} -> \code{Column}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{ColumnSchema}}.}

\item{\code{type}}{of type String inherited from super class \code{\link{ColumnSchema}}.}

\item{\code{nRows}}{of type int inherited from super class \code{\link{ColumnSchema}}.}

\item{\code{size}}{of type int inherited from super class \code{\link{ColumnSchema}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{values}}{of type dynamic.}

\item{\code{metaData}}{object of class \code{\link{ColumnSchemaMetaData}} inherited from super class \code{\link{ColumnSchema}}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{ColumnSchema}}.}

\item{\code{type}}{of type String inherited from super class \code{\link{ColumnSchema}}.}

\item{\code{nRows}}{of type int inherited from super class \code{\link{ColumnSchema}}.}

\item{\code{size}}{of type int inherited from super class \code{\link{ColumnSchema}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{metaData}}{object of class \code{\link{ColumnSchemaMetaData}} inherited from super class \code{\link{ColumnSchema}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Column$new()}}
\item \href{#method-init}{\code{Column$init()}}
\item \href{#method-initJson}{\code{Column$initJson()}}
\item \href{#method-toTson}{\code{Column$toTson()}}
\item \href{#method-print}{\code{Column$print()}}
\item \href{#method-clone}{\code{Column$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Column$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Column$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Column$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Column$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Column$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Column$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
