% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_schema_meta_data.R
\name{ColumnSchemaMetaData}
\alias{ColumnSchemaMetaData}
\title{ColumnSchemaMetaData}
\format{
\code{\link{R6Class}} object.
}
\description{
ColumnSchemaMetaData

ColumnSchemaMetaData
}
\section{Super class}{
\code{\link[tercen:Base]{tercen::Base}} -> \code{ColumnSchemaMetaData}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sort}}{list of type String.}

\item{\code{ascending}}{of type bool.}

\item{\code{quartiles}}{list of type String.}

\item{\code{properties}}{list of class \code{\link{Pair}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ColumnSchemaMetaData$new()}}
\item \href{#method-init}{\code{ColumnSchemaMetaData$init()}}
\item \href{#method-initJson}{\code{ColumnSchemaMetaData$initJson()}}
\item \href{#method-toTson}{\code{ColumnSchemaMetaData$toTson()}}
\item \href{#method-print}{\code{ColumnSchemaMetaData$print()}}
\item \href{#method-clone}{\code{ColumnSchemaMetaData$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnSchemaMetaData$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnSchemaMetaData$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnSchemaMetaData$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnSchemaMetaData$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnSchemaMetaData$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnSchemaMetaData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
