% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computed_table_schema.R
\name{ComputedTableSchema}
\alias{ComputedTableSchema}
\title{ComputedTableSchema}
\format{
\code{\link{R6Class}} object, super class \code{\link{Schema}}.
}
\description{
ComputedTableSchema

ComputedTableSchema
}
\section{Super classes}{
\code{\link[tercenApi:Base]{tercenApi::Base}} -> \code{\link[tercenApi:IdObject]{tercenApi::IdObject}} -> \code{\link[tercenApi:PersistentObject]{tercenApi::PersistentObject}} -> \code{\link[tercenApi:Document]{tercenApi::Document}} -> \code{\link[tercenApi:ProjectDocument]{tercenApi::ProjectDocument}} -> \code{\link[tercenApi:Schema]{tercenApi::Schema}} -> \code{ComputedTableSchema}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nRows}}{of type int inherited from super class \code{\link{Schema}}.}

\item{\code{dataDirectory}}{of type String inherited from super class \code{\link{Schema}}.}

\item{\code{projectId}}{of type String inherited from super class \code{\link{ProjectDocument}}.}

\item{\code{folderId}}{of type String inherited from super class \code{\link{ProjectDocument}}.}

\item{\code{description}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{createdBy}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{tags}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{version}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{authors}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{isPublic}}{of type bool inherited from super class \code{\link{Document}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{columns}}{list of class \code{\link{ColumnSchema}} inherited from super class \code{\link{Schema}}.}

\item{\code{relation}}{object of class \code{\link{Relation}} inherited from super class \code{\link{Schema}}.}

\item{\code{acl}}{object of class \code{\link{Acl}} inherited from super class \code{\link{Document}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{urls}}{list of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}

\item{\code{meta}}{list of class \code{\link{Pair}} inherited from super class \code{\link{Document}}.}

\item{\code{url}}{object of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}

\item{\code{query}}{object of class \code{\link{CubeQuery}}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{nRows}}{of type int inherited from super class \code{\link{Schema}}.}

\item{\code{dataDirectory}}{of type String inherited from super class \code{\link{Schema}}.}

\item{\code{projectId}}{of type String inherited from super class \code{\link{ProjectDocument}}.}

\item{\code{folderId}}{of type String inherited from super class \code{\link{ProjectDocument}}.}

\item{\code{description}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{createdBy}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{tags}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{version}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{authors}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{isPublic}}{of type bool inherited from super class \code{\link{Document}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{columns}}{list of class \code{\link{ColumnSchema}} inherited from super class \code{\link{Schema}}.}

\item{\code{relation}}{object of class \code{\link{Relation}} inherited from super class \code{\link{Schema}}.}

\item{\code{acl}}{object of class \code{\link{Acl}} inherited from super class \code{\link{Document}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{urls}}{list of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}

\item{\code{meta}}{list of class \code{\link{Pair}} inherited from super class \code{\link{Document}}.}

\item{\code{url}}{object of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ComputedTableSchema$new()}}
\item \href{#method-init}{\code{ComputedTableSchema$init()}}
\item \href{#method-initJson}{\code{ComputedTableSchema$initJson()}}
\item \href{#method-toTson}{\code{ComputedTableSchema$toTson()}}
\item \href{#method-clone}{\code{ComputedTableSchema$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="Base" data-id="print">}\href{../../tercenApi/html/Base.html#method-print}{\code{tercenApi::Base$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedTableSchema$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedTableSchema$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedTableSchema$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedTableSchema$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComputedTableSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
