% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_library_service.R
\name{CranLibraryService}
\alias{CranLibraryService}
\title{CranLibraryService}
\format{
\code{\link{R6Class}} object.
}
\description{
CranLibraryService

CranLibraryService
}
\section{Methods}{

\describe{
   \item{\code{packagesGz(repoName)}}{method}
   \item{\code{packagesRds(repoName)}}{method}
   \item{\code{packages(repoName)}}{method}
   \item{\code{archive(repoName,package,filename)}}{method}
   \item{\code{package(repoName,package)}}{method}
}
}

\section{Super class}{
\code{\link[tercenApi:HttpClientService]{tercenApi::HttpClientService}} -> \code{CranLibraryService}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CranLibraryService$new()}}
\item \href{#method-findByOwnerNameVersion}{\code{CranLibraryService$findByOwnerNameVersion()}}
\item \href{#method-packagesGz}{\code{CranLibraryService$packagesGz()}}
\item \href{#method-packagesRds}{\code{CranLibraryService$packagesRds()}}
\item \href{#method-packages}{\code{CranLibraryService$packages()}}
\item \href{#method-archive}{\code{CranLibraryService$archive()}}
\item \href{#method-package}{\code{CranLibraryService$package()}}
\item \href{#method-clone}{\code{CranLibraryService$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="create">}\href{../../tercenApi/html/HttpClientService.html#method-create}{\code{tercenApi::HttpClientService$create()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="delete">}\href{../../tercenApi/html/HttpClientService.html#method-delete}{\code{tercenApi::HttpClientService$delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="findKeys">}\href{../../tercenApi/html/HttpClientService.html#method-findKeys}{\code{tercenApi::HttpClientService$findKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="findStartKeys">}\href{../../tercenApi/html/HttpClientService.html#method-findStartKeys}{\code{tercenApi::HttpClientService$findStartKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="fromTson">}\href{../../tercenApi/html/HttpClientService.html#method-fromTson}{\code{tercenApi::HttpClientService$fromTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="get">}\href{../../tercenApi/html/HttpClientService.html#method-get}{\code{tercenApi::HttpClientService$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="getServiceUri">}\href{../../tercenApi/html/HttpClientService.html#method-getServiceUri}{\code{tercenApi::HttpClientService$getServiceUri()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="list">}\href{../../tercenApi/html/HttpClientService.html#method-list}{\code{tercenApi::HttpClientService$list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="onResponseError">}\href{../../tercenApi/html/HttpClientService.html#method-onResponseError}{\code{tercenApi::HttpClientService$onResponseError()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="toTson">}\href{../../tercenApi/html/HttpClientService.html#method-toTson}{\code{tercenApi::HttpClientService$toTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="update">}\href{../../tercenApi/html/HttpClientService.html#method-update}{\code{tercenApi::HttpClientService$update()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$new(baseRestUri, client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findByOwnerNameVersion"></a>}}
\if{latex}{\out{\hypertarget{method-findByOwnerNameVersion}{}}}
\subsection{Method \code{findByOwnerNameVersion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$findByOwnerNameVersion(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-packagesGz"></a>}}
\if{latex}{\out{\hypertarget{method-packagesGz}{}}}
\subsection{Method \code{packagesGz()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$packagesGz(repoName)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-packagesRds"></a>}}
\if{latex}{\out{\hypertarget{method-packagesRds}{}}}
\subsection{Method \code{packagesRds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$packagesRds(repoName)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-packages"></a>}}
\if{latex}{\out{\hypertarget{method-packages}{}}}
\subsection{Method \code{packages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$packages(repoName)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-archive"></a>}}
\if{latex}{\out{\hypertarget{method-archive}{}}}
\subsection{Method \code{archive()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$archive(repoName, package, filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-package"></a>}}
\if{latex}{\out{\hypertarget{method-package}{}}}
\subsection{Method \code{package()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$package(repoName, package)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
