% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_tab_step.R
\name{CrossTabStep}
\alias{CrossTabStep}
\title{CrossTabStep}
\format{
\code{\link{R6Class}} object, super class \code{\link{NamespaceStep}}, sub classes \code{\link{DataStep}}.
}
\description{
CrossTabStep

CrossTabStep
}
\section{Super classes}{
\code{\link[tercenApi:Base]{tercenApi::Base}} -> \code{\link[tercenApi:IdObject]{tercenApi::IdObject}} -> \code{\link[tercenApi:Step]{tercenApi::Step}} -> \code{\link[tercenApi:ModelStep]{tercenApi::ModelStep}} -> \code{\link[tercenApi:RelationStep]{tercenApi::RelationStep}} -> \code{\link[tercenApi:NamespaceStep]{tercenApi::NamespaceStep}} -> \code{CrossTabStep}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{groupId}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{inputs}}{list of class \code{\link{InputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{outputs}}{list of class \code{\link{OutputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{rectangle}}{object of class \code{\link{Rectangle}} inherited from super class \code{\link{Step}}.}

\item{\code{state}}{object of class \code{\link{StepState}} inherited from super class \code{\link{Step}}.}

\item{\code{model}}{object of class \code{\link{Crosstab}}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{groupId}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{inputs}}{list of class \code{\link{InputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{outputs}}{list of class \code{\link{OutputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{rectangle}}{object of class \code{\link{Rectangle}} inherited from super class \code{\link{Step}}.}

\item{\code{state}}{object of class \code{\link{StepState}} inherited from super class \code{\link{Step}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CrossTabStep$new()}}
\item \href{#method-init}{\code{CrossTabStep$init()}}
\item \href{#method-initJson}{\code{CrossTabStep$initJson()}}
\item \href{#method-toTson}{\code{CrossTabStep$toTson()}}
\item \href{#method-clone}{\code{CrossTabStep$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="Base" data-id="print">}\href{../../tercenApi/html/Base.html#method-print}{\code{tercenApi::Base$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossTabStep$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossTabStep$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossTabStep$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossTabStep$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossTabStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
