% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_s_v_file_metadata.R
\name{CSVFileMetadata}
\alias{CSVFileMetadata}
\title{CSVFileMetadata}
\format{
\code{\link{R6Class}} object, super class \code{\link{FileMetadata}}.
}
\description{
CSVFileMetadata

CSVFileMetadata
}
\section{Super classes}{
\code{\link[tercenApi:Base]{tercenApi::Base}} -> \code{\link[tercenApi:FileMetadata]{tercenApi::FileMetadata}} -> \code{CSVFileMetadata}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{contentType}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{cacheControl}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{contentEncoding}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{contentLanguage}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{md5Hash}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{separator}}{of type String.}

\item{\code{quote}}{of type String.}

\item{\code{headers}}{of type bool.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{contentType}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{cacheControl}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{contentEncoding}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{contentLanguage}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{md5Hash}}{of type String inherited from super class \code{\link{FileMetadata}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CSVFileMetadata$new()}}
\item \href{#method-init}{\code{CSVFileMetadata$init()}}
\item \href{#method-initJson}{\code{CSVFileMetadata$initJson()}}
\item \href{#method-toTson}{\code{CSVFileMetadata$toTson()}}
\item \href{#method-clone}{\code{CSVFileMetadata$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="Base" data-id="print">}\href{../../tercenApi/html/Base.html#method-print}{\code{tercenApi::Base$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVFileMetadata$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVFileMetadata$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVFileMetadata$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVFileMetadata$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVFileMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
