% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_axis_query.R
\name{CubeAxisQuery}
\alias{CubeAxisQuery}
\title{CubeAxisQuery}
\format{
\code{\link{R6Class}} object.
}
\description{
CubeAxisQuery

CubeAxisQuery
}
\section{Super class}{
\code{\link[tercen:Base]{tercen::Base}} -> \code{CubeAxisQuery}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pointSize}}{of type int.}

\item{\code{chartType}}{of type String.}

\item{\code{yAxis}}{object of class \code{\link{Factor}}.}

\item{\code{yAxisSettings}}{object of class \code{\link{AxisSettings}}.}

\item{\code{xAxis}}{object of class \code{\link{Factor}}.}

\item{\code{xAxisSettings}}{object of class \code{\link{AxisSettings}}.}

\item{\code{errors}}{list of class \code{\link{Factor}}.}

\item{\code{labels}}{list of class \code{\link{Factor}}.}

\item{\code{colors}}{list of class \code{\link{Factor}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CubeAxisQuery$new()}}
\item \href{#method-init}{\code{CubeAxisQuery$init()}}
\item \href{#method-initJson}{\code{CubeAxisQuery$initJson()}}
\item \href{#method-toTson}{\code{CubeAxisQuery$toTson()}}
\item \href{#method-print}{\code{CubeAxisQuery$print()}}
\item \href{#method-clone}{\code{CubeAxisQuery$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeAxisQuery$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeAxisQuery$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeAxisQuery$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeAxisQuery$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeAxisQuery$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeAxisQuery$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
