% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_query.R
\name{CubeQuery}
\alias{CubeQuery}
\title{CubeQuery}
\format{
\code{\link{R6Class}} object.
}
\description{
CubeQuery

CubeQuery
}
\section{Super class}{
\code{\link[tercenApi:Base]{tercenApi::Base}} -> \code{CubeQuery}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{qtHash}}{of type String.}

\item{\code{columnHash}}{of type String.}

\item{\code{rowHash}}{of type String.}

\item{\code{relation}}{object of class \code{\link{Relation}}.}

\item{\code{colColumns}}{list of class \code{\link{Factor}}.}

\item{\code{rowColumns}}{list of class \code{\link{Factor}}.}

\item{\code{axisQueries}}{list of class \code{\link{CubeAxisQuery}}.}

\item{\code{filters}}{object of class \code{\link{Filters}}.}

\item{\code{operatorSettings}}{object of class \code{\link{OperatorSettings}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CubeQuery$new()}}
\item \href{#method-init}{\code{CubeQuery$init()}}
\item \href{#method-initJson}{\code{CubeQuery$initJson()}}
\item \href{#method-toTson}{\code{CubeQuery$toTson()}}
\item \href{#method-clone}{\code{CubeQuery$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="Base" data-id="print">}\href{../../tercenApi/html/Base.html#method-print}{\code{tercenApi::Base$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeQuery$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeQuery$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeQuery$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeQuery$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeQuery$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
