% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_step.R
\name{DataStep}
\alias{DataStep}
\title{DataStep}
\format{
\code{\link{R6Class}} object, super class \code{\link{CrossTabStep}}.
}
\description{
DataStep

DataStep
}
\section{Super classes}{
\code{\link[tercenApi:Base]{tercenApi::Base}} -> \code{\link[tercenApi:IdObject]{tercenApi::IdObject}} -> \code{\link[tercenApi:Step]{tercenApi::Step}} -> \code{\link[tercenApi:ModelStep]{tercenApi::ModelStep}} -> \code{\link[tercenApi:RelationStep]{tercenApi::RelationStep}} -> \code{\link[tercenApi:NamespaceStep]{tercenApi::NamespaceStep}} -> \code{\link[tercenApi:CrossTabStep]{tercenApi::CrossTabStep}} -> \code{DataStep}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{groupId}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{model}}{object of class \code{\link{Crosstab}} inherited from super class \code{\link{CrossTabStep}}.}

\item{\code{inputs}}{list of class \code{\link{InputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{outputs}}{list of class \code{\link{OutputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{rectangle}}{object of class \code{\link{Rectangle}} inherited from super class \code{\link{Step}}.}

\item{\code{state}}{object of class \code{\link{StepState}} inherited from super class \code{\link{Step}}.}

\item{\code{computedRelation}}{object of class \code{\link{Relation}}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{groupId}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{model}}{object of class \code{\link{Crosstab}} inherited from super class \code{\link{CrossTabStep}}.}

\item{\code{inputs}}{list of class \code{\link{InputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{outputs}}{list of class \code{\link{OutputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{rectangle}}{object of class \code{\link{Rectangle}} inherited from super class \code{\link{Step}}.}

\item{\code{state}}{object of class \code{\link{StepState}} inherited from super class \code{\link{Step}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataStep$new()}}
\item \href{#method-init}{\code{DataStep$init()}}
\item \href{#method-initJson}{\code{DataStep$initJson()}}
\item \href{#method-toTson}{\code{DataStep$toTson()}}
\item \href{#method-print}{\code{DataStep$print()}}
\item \href{#method-clone}{\code{DataStep$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStep$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStep$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStep$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStep$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStep$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
