% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_document.R
\name{FolderDocument}
\alias{FolderDocument}
\title{FolderDocument}
\format{
\code{\link{R6Class}} object, super class \code{\link{ProjectDocument}}.
}
\description{
FolderDocument

FolderDocument
}
\section{Super classes}{
\code{\link[tercen:Base]{tercen::Base}} -> \code{\link[tercen:IdObject]{tercen::IdObject}} -> \code{\link[tercen:PersistentObject]{tercen::PersistentObject}} -> \code{\link[tercen:Document]{tercen::Document}} -> \code{\link[tercen:ProjectDocument]{tercen::ProjectDocument}} -> \code{FolderDocument}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{projectId}}{of type String inherited from super class \code{\link{ProjectDocument}}.}

\item{\code{folderId}}{of type String inherited from super class \code{\link{ProjectDocument}}.}

\item{\code{description}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{createdBy}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{tags}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{version}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{authors}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{isPublic}}{of type bool inherited from super class \code{\link{Document}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{acl}}{object of class \code{\link{Acl}} inherited from super class \code{\link{Document}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{urls}}{list of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}

\item{\code{meta}}{list of class \code{\link{Pair}} inherited from super class \code{\link{Document}}.}

\item{\code{url}}{object of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{projectId}}{of type String inherited from super class \code{\link{ProjectDocument}}.}

\item{\code{folderId}}{of type String inherited from super class \code{\link{ProjectDocument}}.}

\item{\code{description}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{createdBy}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{tags}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{version}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{authors}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{isPublic}}{of type bool inherited from super class \code{\link{Document}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{acl}}{object of class \code{\link{Acl}} inherited from super class \code{\link{Document}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{urls}}{list of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}

\item{\code{meta}}{list of class \code{\link{Pair}} inherited from super class \code{\link{Document}}.}

\item{\code{url}}{object of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FolderDocument$new()}}
\item \href{#method-init}{\code{FolderDocument$init()}}
\item \href{#method-initJson}{\code{FolderDocument$initJson()}}
\item \href{#method-toTson}{\code{FolderDocument$toTson()}}
\item \href{#method-print}{\code{FolderDocument$print()}}
\item \href{#method-clone}{\code{FolderDocument$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderDocument$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderDocument$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderDocument$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderDocument$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderDocument$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderDocument$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
