% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_step.R
\name{GroupStep}
\alias{GroupStep}
\title{GroupStep}
\format{
\code{\link{R6Class}} object, super class \code{\link{RelationStep}}.
}
\description{
GroupStep

GroupStep
}
\section{Super classes}{
\code{\link[tercen:Base]{tercen::Base}} -> \code{\link[tercen:IdObject]{tercen::IdObject}} -> \code{\link[tercen:Step]{tercen::Step}} -> \code{\link[tercen:ModelStep]{tercen::ModelStep}} -> \code{\link[tercen:RelationStep]{tercen::RelationStep}} -> \code{GroupStep}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{groupId}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{appId}}{of type String.}

\item{\code{appName}}{of type String.}

\item{\code{version}}{of type String.}

\item{\code{inputs}}{list of class \code{\link{InputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{outputs}}{list of class \code{\link{OutputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{rectangle}}{object of class \code{\link{Rectangle}} inherited from super class \code{\link{Step}}.}

\item{\code{state}}{object of class \code{\link{StepState}} inherited from super class \code{\link{Step}}.}

\item{\code{offset}}{object of class \code{\link{Point}}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{groupId}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{inputs}}{list of class \code{\link{InputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{outputs}}{list of class \code{\link{OutputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{rectangle}}{object of class \code{\link{Rectangle}} inherited from super class \code{\link{Step}}.}

\item{\code{state}}{object of class \code{\link{StepState}} inherited from super class \code{\link{Step}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GroupStep$new()}}
\item \href{#method-init}{\code{GroupStep$init()}}
\item \href{#method-initJson}{\code{GroupStep$initJson()}}
\item \href{#method-toTson}{\code{GroupStep$toTson()}}
\item \href{#method-print}{\code{GroupStep$print()}}
\item \href{#method-clone}{\code{GroupStep$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupStep$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupStep$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupStep$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupStep$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupStep$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
