% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa-tercen.R
\name{HttpClientService}
\alias{HttpClientService}
\title{HttpClientService}
\description{
HttpClientService

HttpClientService
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HttpClientService$new()}}
\item \href{#method-onResponseError}{\code{HttpClientService$onResponseError()}}
\item \href{#method-toTson}{\code{HttpClientService$toTson()}}
\item \href{#method-fromTson}{\code{HttpClientService$fromTson()}}
\item \href{#method-getServiceUri}{\code{HttpClientService$getServiceUri()}}
\item \href{#method-create}{\code{HttpClientService$create()}}
\item \href{#method-get}{\code{HttpClientService$get()}}
\item \href{#method-delete}{\code{HttpClientService$delete()}}
\item \href{#method-update}{\code{HttpClientService$update()}}
\item \href{#method-list}{\code{HttpClientService$list()}}
\item \href{#method-findStartKeys}{\code{HttpClientService$findStartKeys()}}
\item \href{#method-findKeys}{\code{HttpClientService$findKeys()}}
\item \href{#method-clone}{\code{HttpClientService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientService$new(baseRestUri, client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-onResponseError"></a>}}
\if{latex}{\out{\hypertarget{method-onResponseError}{}}}
\subsection{Method \code{onResponseError()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientService$onResponseError(response, msg = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientService$toTson(object)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromTson"></a>}}
\if{latex}{\out{\hypertarget{method-fromTson}{}}}
\subsection{Method \code{fromTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientService$fromTson(object)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getServiceUri"></a>}}
\if{latex}{\out{\hypertarget{method-getServiceUri}{}}}
\subsection{Method \code{getServiceUri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientService$getServiceUri(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create"></a>}}
\if{latex}{\out{\hypertarget{method-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientService$create(object)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientService$get(id, useFactory = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientService$delete(id, rev)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientService$update(object)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list"></a>}}
\if{latex}{\out{\hypertarget{method-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientService$list(ids, useFactory = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findStartKeys"></a>}}
\if{latex}{\out{\hypertarget{method-findStartKeys}{}}}
\subsection{Method \code{findStartKeys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientService$findStartKeys(
  viewName,
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findKeys"></a>}}
\if{latex}{\out{\hypertarget{method-findKeys}{}}}
\subsection{Method \code{findKeys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientService$findKeys(viewName, keys = NULL, useFactory = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
