% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet_palette.R
\name{JetPalette}
\alias{JetPalette}
\title{JetPalette}
\format{
\code{\link{R6Class}} object, super class \code{\link{RampPalette}}.
}
\description{
JetPalette

JetPalette
}
\section{Super classes}{
\code{\link[tercen:Base]{tercen::Base}} -> \code{\link[tercen:Palette]{tercen::Palette}} -> \code{\link[tercen:RampPalette]{tercen::RampPalette}} -> \code{JetPalette}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isUserDefined}}{of type bool inherited from super class \code{\link{RampPalette}}.}

\item{\code{backcolor}}{of type int inherited from super class \code{\link{Palette}}.}

\item{\code{doubleColorElements}}{list of class \code{\link{DoubleColorElement}} inherited from super class \code{\link{RampPalette}}.}

\item{\code{properties}}{list of class \code{\link{PropertyValue}} inherited from super class \code{\link{Palette}}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{isUserDefined}}{of type bool inherited from super class \code{\link{RampPalette}}.}

\item{\code{backcolor}}{of type int inherited from super class \code{\link{Palette}}.}

\item{\code{doubleColorElements}}{list of class \code{\link{DoubleColorElement}} inherited from super class \code{\link{RampPalette}}.}

\item{\code{properties}}{list of class \code{\link{PropertyValue}} inherited from super class \code{\link{Palette}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JetPalette$new()}}
\item \href{#method-init}{\code{JetPalette$init()}}
\item \href{#method-initJson}{\code{JetPalette$initJson()}}
\item \href{#method-toTson}{\code{JetPalette$toTson()}}
\item \href{#method-print}{\code{JetPalette$print()}}
\item \href{#method-clone}{\code{JetPalette$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JetPalette$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JetPalette$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JetPalette$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JetPalette$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JetPalette$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JetPalette$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
