% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_filter.R
\name{NamedFilter}
\alias{NamedFilter}
\title{NamedFilter}
\format{
\code{\link{R6Class}} object, super class \code{\link{Filter}}.
}
\description{
NamedFilter

NamedFilter
}
\section{Super classes}{
\code{\link[tercen:Base]{tercen::Base}} -> \code{\link[tercen:FilterTopExpr]{tercen::FilterTopExpr}} -> \code{\link[tercen:Filter]{tercen::Filter}} -> \code{NamedFilter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{logical}}{of type String inherited from super class \code{\link{Filter}}.}

\item{\code{not}}{of type bool inherited from super class \code{\link{Filter}}.}

\item{\code{name}}{of type String.}

\item{\code{filterExprs}}{list of class \code{\link{FilterTopExpr}} inherited from super class \code{\link{Filter}}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{logical}}{of type String inherited from super class \code{\link{Filter}}.}

\item{\code{not}}{of type bool inherited from super class \code{\link{Filter}}.}

\item{\code{filterExprs}}{list of class \code{\link{FilterTopExpr}} inherited from super class \code{\link{Filter}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NamedFilter$new()}}
\item \href{#method-init}{\code{NamedFilter$init()}}
\item \href{#method-initJson}{\code{NamedFilter$initJson()}}
\item \href{#method-toTson}{\code{NamedFilter$toTson()}}
\item \href{#method-print}{\code{NamedFilter$print()}}
\item \href{#method-clone}{\code{NamedFilter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedFilter$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedFilter$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedFilter$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedFilter$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedFilter$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamedFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
