% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_service.R
\name{ProjectService}
\alias{ProjectService}
\title{ProjectService}
\format{
\code{\link{R6Class}} object.
}
\description{
ProjectService

ProjectService
}
\section{Methods}{

\describe{
   \item{\code{profiles(projectId)}}{method}
   \item{\code{resourceSummary(projectId)}}{method}
   \item{\code{explore(category,start,limit)}}{method}
   \item{\code{recentProjects(userId)}}{method}
   \item{\code{cloneProject(projectId,project)}}{method}
}
}

\section{Super class}{
\code{\link[tercen:HttpClientService]{tercen::HttpClientService}} -> \code{ProjectService}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectService$new()}}
\item \href{#method-findByIsPublicAndLastModifiedDate}{\code{ProjectService$findByIsPublicAndLastModifiedDate()}}
\item \href{#method-findByTeamAndIsPublicAndLastModifiedDate}{\code{ProjectService$findByTeamAndIsPublicAndLastModifiedDate()}}
\item \href{#method-profiles}{\code{ProjectService$profiles()}}
\item \href{#method-resourceSummary}{\code{ProjectService$resourceSummary()}}
\item \href{#method-explore}{\code{ProjectService$explore()}}
\item \href{#method-recentProjects}{\code{ProjectService$recentProjects()}}
\item \href{#method-cloneProject}{\code{ProjectService$cloneProject()}}
\item \href{#method-clone}{\code{ProjectService$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="create">}\href{../../tercen/html/HttpClientService.html#method-create}{\code{tercen::HttpClientService$create()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="delete">}\href{../../tercen/html/HttpClientService.html#method-delete}{\code{tercen::HttpClientService$delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="findKeys">}\href{../../tercen/html/HttpClientService.html#method-findKeys}{\code{tercen::HttpClientService$findKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="findStartKeys">}\href{../../tercen/html/HttpClientService.html#method-findStartKeys}{\code{tercen::HttpClientService$findStartKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="fromTson">}\href{../../tercen/html/HttpClientService.html#method-fromTson}{\code{tercen::HttpClientService$fromTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="get">}\href{../../tercen/html/HttpClientService.html#method-get}{\code{tercen::HttpClientService$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="getServiceUri">}\href{../../tercen/html/HttpClientService.html#method-getServiceUri}{\code{tercen::HttpClientService$getServiceUri()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="list">}\href{../../tercen/html/HttpClientService.html#method-list}{\code{tercen::HttpClientService$list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="onResponseError">}\href{../../tercen/html/HttpClientService.html#method-onResponseError}{\code{tercen::HttpClientService$onResponseError()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="toTson">}\href{../../tercen/html/HttpClientService.html#method-toTson}{\code{tercen::HttpClientService$toTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="update">}\href{../../tercen/html/HttpClientService.html#method-update}{\code{tercen::HttpClientService$update()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectService$new(baseRestUri, client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findByIsPublicAndLastModifiedDate"></a>}}
\if{latex}{\out{\hypertarget{method-findByIsPublicAndLastModifiedDate}{}}}
\subsection{Method \code{findByIsPublicAndLastModifiedDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectService$findByIsPublicAndLastModifiedDate(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findByTeamAndIsPublicAndLastModifiedDate"></a>}}
\if{latex}{\out{\hypertarget{method-findByTeamAndIsPublicAndLastModifiedDate}{}}}
\subsection{Method \code{findByTeamAndIsPublicAndLastModifiedDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectService$findByTeamAndIsPublicAndLastModifiedDate(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-profiles"></a>}}
\if{latex}{\out{\hypertarget{method-profiles}{}}}
\subsection{Method \code{profiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectService$profiles(projectId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resourceSummary"></a>}}
\if{latex}{\out{\hypertarget{method-resourceSummary}{}}}
\subsection{Method \code{resourceSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectService$resourceSummary(projectId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-explore"></a>}}
\if{latex}{\out{\hypertarget{method-explore}{}}}
\subsection{Method \code{explore()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectService$explore(category, start, limit)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recentProjects"></a>}}
\if{latex}{\out{\hypertarget{method-recentProjects}{}}}
\subsection{Method \code{recentProjects()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectService$recentProjects(userId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cloneProject"></a>}}
\if{latex}{\out{\hypertarget{method-cloneProject}{}}}
\subsection{Method \code{cloneProject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectService$cloneProject(projectId, project)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
