% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_computation_task.R
\name{RunComputationTask}
\alias{RunComputationTask}
\title{RunComputationTask}
\format{
\code{\link{R6Class}} object, super class \code{\link{SaveComputationResultTask}}.
}
\description{
RunComputationTask

RunComputationTask
}
\section{Super classes}{
\code{\link[tercenApi:Base]{tercenApi::Base}} -> \code{\link[tercenApi:IdObject]{tercenApi::IdObject}} -> \code{\link[tercenApi:PersistentObject]{tercenApi::PersistentObject}} -> \code{\link[tercenApi:Task]{tercenApi::Task}} -> \code{\link[tercenApi:ProjectTask]{tercenApi::ProjectTask}} -> \code{\link[tercenApi:CubeQueryTask]{tercenApi::CubeQueryTask}} -> \code{\link[tercenApi:ComputationTask]{tercenApi::ComputationTask}} -> \code{\link[tercenApi:SaveComputationResultTask]{tercenApi::SaveComputationResultTask}} -> \code{RunComputationTask}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parentTaskId}}{of type String inherited from super class \code{\link{ComputationTask}}.}

\item{\code{fileResultId}}{of type String inherited from super class \code{\link{ComputationTask}}.}

\item{\code{removeOnGC}}{of type bool inherited from super class \code{\link{CubeQueryTask}}.}

\item{\code{schemaIds}}{list of type String inherited from super class \code{\link{CubeQueryTask}}.}

\item{\code{projectId}}{of type String inherited from super class \code{\link{ProjectTask}}.}

\item{\code{duration}}{of type double inherited from super class \code{\link{Task}}.}

\item{\code{owner}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{taskHash}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{channelId}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{computedRelation}}{object of class \code{\link{Relation}} inherited from super class \code{\link{ComputationTask}}.}

\item{\code{query}}{object of class \code{\link{CubeQuery}} inherited from super class \code{\link{CubeQueryTask}}.}

\item{\code{environment}}{list of class \code{\link{Pair}} inherited from super class \code{\link{Task}}.}

\item{\code{state}}{object of class \code{\link{State}} inherited from super class \code{\link{Task}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{runDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{completedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{aclContext}}{object of class \code{\link{AclContext}} inherited from super class \code{\link{Task}}.}

\item{\code{meta}}{list of class \code{\link{Pair}} inherited from super class \code{\link{Task}}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parentTaskId}}{of type String inherited from super class \code{\link{ComputationTask}}.}

\item{\code{fileResultId}}{of type String inherited from super class \code{\link{ComputationTask}}.}

\item{\code{removeOnGC}}{of type bool inherited from super class \code{\link{CubeQueryTask}}.}

\item{\code{schemaIds}}{list of type String inherited from super class \code{\link{CubeQueryTask}}.}

\item{\code{projectId}}{of type String inherited from super class \code{\link{ProjectTask}}.}

\item{\code{duration}}{of type double inherited from super class \code{\link{Task}}.}

\item{\code{owner}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{taskHash}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{channelId}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{computedRelation}}{object of class \code{\link{Relation}} inherited from super class \code{\link{ComputationTask}}.}

\item{\code{query}}{object of class \code{\link{CubeQuery}} inherited from super class \code{\link{CubeQueryTask}}.}

\item{\code{environment}}{list of class \code{\link{Pair}} inherited from super class \code{\link{Task}}.}

\item{\code{state}}{object of class \code{\link{State}} inherited from super class \code{\link{Task}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{runDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{completedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{aclContext}}{object of class \code{\link{AclContext}} inherited from super class \code{\link{Task}}.}

\item{\code{meta}}{list of class \code{\link{Pair}} inherited from super class \code{\link{Task}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RunComputationTask$new()}}
\item \href{#method-init}{\code{RunComputationTask$init()}}
\item \href{#method-initJson}{\code{RunComputationTask$initJson()}}
\item \href{#method-toTson}{\code{RunComputationTask$toTson()}}
\item \href{#method-clone}{\code{RunComputationTask$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="Base" data-id="print">}\href{../../tercenApi/html/Base.html#method-print}{\code{tercenApi::Base$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunComputationTask$new(json = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunComputationTask$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initJson"></a>}}
\if{latex}{\out{\hypertarget{method-initJson}{}}}
\subsection{Method \code{initJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunComputationTask$initJson(json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toTson"></a>}}
\if{latex}{\out{\hypertarget{method-toTson}{}}}
\subsection{Method \code{toTson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunComputationTask$toTson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunComputationTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
