% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription_plan_service.R
\name{SubscriptionPlanService}
\alias{SubscriptionPlanService}
\title{SubscriptionPlanService}
\format{
\code{\link{R6Class}} object.
}
\description{
SubscriptionPlanService

SubscriptionPlanService
}
\section{Methods}{

\describe{
   \item{\code{getSubscriptionPlans(userId)}}{method}
   \item{\code{getPlans(userId)}}{method}
   \item{\code{createSubscriptionPlan(userId,plan,successUrl,cancelUrl)}}{method}
   \item{\code{setSubscriptionPlanStatus(subscriptionPlanId,status)}}{method}
   \item{\code{updatePaymentMethod(subscriptionPlanId,successUrl,cancelUrl)}}{method}
   \item{\code{setUpdatePaymentMethodStatus(subscriptionPlanId,status)}}{method}
   \item{\code{cancelSubscription(subscriptionPlanId)}}{method}
   \item{\code{upgradeSubscription(subscriptionPlanId,plan)}}{method}
}
}

\section{Super class}{
\code{\link[tercen:HttpClientService]{tercen::HttpClientService}} -> \code{SubscriptionPlanService}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubscriptionPlanService$new()}}
\item \href{#method-findByOwner}{\code{SubscriptionPlanService$findByOwner()}}
\item \href{#method-findSubscriptionPlanByCheckoutSessionId}{\code{SubscriptionPlanService$findSubscriptionPlanByCheckoutSessionId()}}
\item \href{#method-getSubscriptionPlans}{\code{SubscriptionPlanService$getSubscriptionPlans()}}
\item \href{#method-getPlans}{\code{SubscriptionPlanService$getPlans()}}
\item \href{#method-createSubscriptionPlan}{\code{SubscriptionPlanService$createSubscriptionPlan()}}
\item \href{#method-setSubscriptionPlanStatus}{\code{SubscriptionPlanService$setSubscriptionPlanStatus()}}
\item \href{#method-updatePaymentMethod}{\code{SubscriptionPlanService$updatePaymentMethod()}}
\item \href{#method-setUpdatePaymentMethodStatus}{\code{SubscriptionPlanService$setUpdatePaymentMethodStatus()}}
\item \href{#method-cancelSubscription}{\code{SubscriptionPlanService$cancelSubscription()}}
\item \href{#method-upgradeSubscription}{\code{SubscriptionPlanService$upgradeSubscription()}}
\item \href{#method-clone}{\code{SubscriptionPlanService$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="create">}\href{../../tercen/html/HttpClientService.html#method-create}{\code{tercen::HttpClientService$create()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="delete">}\href{../../tercen/html/HttpClientService.html#method-delete}{\code{tercen::HttpClientService$delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="findKeys">}\href{../../tercen/html/HttpClientService.html#method-findKeys}{\code{tercen::HttpClientService$findKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="findStartKeys">}\href{../../tercen/html/HttpClientService.html#method-findStartKeys}{\code{tercen::HttpClientService$findStartKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="fromTson">}\href{../../tercen/html/HttpClientService.html#method-fromTson}{\code{tercen::HttpClientService$fromTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="get">}\href{../../tercen/html/HttpClientService.html#method-get}{\code{tercen::HttpClientService$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="getServiceUri">}\href{../../tercen/html/HttpClientService.html#method-getServiceUri}{\code{tercen::HttpClientService$getServiceUri()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="list">}\href{../../tercen/html/HttpClientService.html#method-list}{\code{tercen::HttpClientService$list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="onResponseError">}\href{../../tercen/html/HttpClientService.html#method-onResponseError}{\code{tercen::HttpClientService$onResponseError()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="toTson">}\href{../../tercen/html/HttpClientService.html#method-toTson}{\code{tercen::HttpClientService$toTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="update">}\href{../../tercen/html/HttpClientService.html#method-update}{\code{tercen::HttpClientService$update()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionPlanService$new(baseRestUri, client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findByOwner"></a>}}
\if{latex}{\out{\hypertarget{method-findByOwner}{}}}
\subsection{Method \code{findByOwner()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionPlanService$findByOwner(keys = NULL, useFactory = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findSubscriptionPlanByCheckoutSessionId"></a>}}
\if{latex}{\out{\hypertarget{method-findSubscriptionPlanByCheckoutSessionId}{}}}
\subsection{Method \code{findSubscriptionPlanByCheckoutSessionId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionPlanService$findSubscriptionPlanByCheckoutSessionId(
  keys = NULL,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSubscriptionPlans"></a>}}
\if{latex}{\out{\hypertarget{method-getSubscriptionPlans}{}}}
\subsection{Method \code{getSubscriptionPlans()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionPlanService$getSubscriptionPlans(userId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPlans"></a>}}
\if{latex}{\out{\hypertarget{method-getPlans}{}}}
\subsection{Method \code{getPlans()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionPlanService$getPlans(userId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createSubscriptionPlan"></a>}}
\if{latex}{\out{\hypertarget{method-createSubscriptionPlan}{}}}
\subsection{Method \code{createSubscriptionPlan()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionPlanService$createSubscriptionPlan(
  userId,
  plan,
  successUrl,
  cancelUrl
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSubscriptionPlanStatus"></a>}}
\if{latex}{\out{\hypertarget{method-setSubscriptionPlanStatus}{}}}
\subsection{Method \code{setSubscriptionPlanStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionPlanService$setSubscriptionPlanStatus(subscriptionPlanId, status)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updatePaymentMethod"></a>}}
\if{latex}{\out{\hypertarget{method-updatePaymentMethod}{}}}
\subsection{Method \code{updatePaymentMethod()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionPlanService$updatePaymentMethod(
  subscriptionPlanId,
  successUrl,
  cancelUrl
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setUpdatePaymentMethodStatus"></a>}}
\if{latex}{\out{\hypertarget{method-setUpdatePaymentMethodStatus}{}}}
\subsection{Method \code{setUpdatePaymentMethodStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionPlanService$setUpdatePaymentMethodStatus(
  subscriptionPlanId,
  status
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cancelSubscription"></a>}}
\if{latex}{\out{\hypertarget{method-cancelSubscription}{}}}
\subsection{Method \code{cancelSubscription()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionPlanService$cancelSubscription(subscriptionPlanId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upgradeSubscription"></a>}}
\if{latex}{\out{\hypertarget{method-upgradeSubscription}{}}}
\subsection{Method \code{upgradeSubscription()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionPlanService$upgradeSubscription(subscriptionPlanId, plan)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubscriptionPlanService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
