% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_service.R
\name{TeamService}
\alias{TeamService}
\title{TeamService}
\format{
\code{\link{R6Class}} object.
}
\description{
TeamService

TeamService
}
\section{Methods}{

\describe{
   \item{\code{profiles(teamId)}}{method}
   \item{\code{resourceSummary(teamId)}}{method}
}
}

\section{Super class}{
\code{\link[tercen:HttpClientService]{tercen::HttpClientService}} -> \code{TeamService}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamService$new()}}
\item \href{#method-findTeamByNameByLastModifiedDate}{\code{TeamService$findTeamByNameByLastModifiedDate()}}
\item \href{#method-findTeamByOwner}{\code{TeamService$findTeamByOwner()}}
\item \href{#method-findTeamByName}{\code{TeamService$findTeamByName()}}
\item \href{#method-profiles}{\code{TeamService$profiles()}}
\item \href{#method-resourceSummary}{\code{TeamService$resourceSummary()}}
\item \href{#method-clone}{\code{TeamService$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="create">}\href{../../tercen/html/HttpClientService.html#method-create}{\code{tercen::HttpClientService$create()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="delete">}\href{../../tercen/html/HttpClientService.html#method-delete}{\code{tercen::HttpClientService$delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="findKeys">}\href{../../tercen/html/HttpClientService.html#method-findKeys}{\code{tercen::HttpClientService$findKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="findStartKeys">}\href{../../tercen/html/HttpClientService.html#method-findStartKeys}{\code{tercen::HttpClientService$findStartKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="fromTson">}\href{../../tercen/html/HttpClientService.html#method-fromTson}{\code{tercen::HttpClientService$fromTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="get">}\href{../../tercen/html/HttpClientService.html#method-get}{\code{tercen::HttpClientService$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="getServiceUri">}\href{../../tercen/html/HttpClientService.html#method-getServiceUri}{\code{tercen::HttpClientService$getServiceUri()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="list">}\href{../../tercen/html/HttpClientService.html#method-list}{\code{tercen::HttpClientService$list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="onResponseError">}\href{../../tercen/html/HttpClientService.html#method-onResponseError}{\code{tercen::HttpClientService$onResponseError()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="toTson">}\href{../../tercen/html/HttpClientService.html#method-toTson}{\code{tercen::HttpClientService$toTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="update">}\href{../../tercen/html/HttpClientService.html#method-update}{\code{tercen::HttpClientService$update()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamService$new(baseRestUri, client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findTeamByNameByLastModifiedDate"></a>}}
\if{latex}{\out{\hypertarget{method-findTeamByNameByLastModifiedDate}{}}}
\subsection{Method \code{findTeamByNameByLastModifiedDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamService$findTeamByNameByLastModifiedDate(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findTeamByOwner"></a>}}
\if{latex}{\out{\hypertarget{method-findTeamByOwner}{}}}
\subsection{Method \code{findTeamByOwner()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamService$findTeamByOwner(keys = NULL, useFactory = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findTeamByName"></a>}}
\if{latex}{\out{\hypertarget{method-findTeamByName}{}}}
\subsection{Method \code{findTeamByName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamService$findTeamByName(keys = NULL, useFactory = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-profiles"></a>}}
\if{latex}{\out{\hypertarget{method-profiles}{}}}
\subsection{Method \code{profiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamService$profiles(teamId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resourceSummary"></a>}}
\if{latex}{\out{\hypertarget{method-resourceSummary}{}}}
\subsection{Method \code{resourceSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamService$resourceSummary(teamId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
