% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_service.R
\name{UserService}
\alias{UserService}
\title{UserService}
\format{
\code{\link{R6Class}} object.
}
\description{
UserService

UserService
}
\section{Methods}{

\describe{
   \item{\code{getSamlMessage(type)}}{method}
   \item{\code{cookieConsent(dummy)}}{method}
   \item{\code{connect(usernameOrEmail,password)}}{method}
   \item{\code{connect2(domain,usernameOrEmail,password)}}{method}
   \item{\code{createUser(user,password)}}{method}
   \item{\code{hasUserName(username)}}{method}
   \item{\code{updatePassword(userId,password)}}{method}
   \item{\code{updateBillingInfo(userId,billingInfo)}}{method}
   \item{\code{viesInfo(country_code,vatNumber)}}{method}
   \item{\code{summary(userId)}}{method}
   \item{\code{resourceSummary(userId)}}{method}
   \item{\code{profiles(userId)}}{method}
   \item{\code{createToken(userId,validityInSeconds)}}{method}
   \item{\code{isTokenValid(token)}}{method}
   \item{\code{setTeamPrivilege(username,principal,privilege)}}{method}
   \item{\code{getServerVersion(module)}}{method}
   \item{\code{getInvited(email)}}{method}
   \item{\code{sendValidationMail(email)}}{method}
   \item{\code{sendResetPasswordEmail(email)}}{method}
   \item{\code{changeUserPassword(token,password)}}{method}
   \item{\code{validateUser(token)}}{method}
   \item{\code{canCreatePrivateProject(teamOrUserId)}}{method}
}
}

\section{Super class}{
\code{\link[tercenApi:HttpClientService]{tercenApi::HttpClientService}} -> \code{UserService}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserService$new()}}
\item \href{#method-findTeamMembers}{\code{UserService$findTeamMembers()}}
\item \href{#method-findUserByCreatedDateAndName}{\code{UserService$findUserByCreatedDateAndName()}}
\item \href{#method-findUserByEmail}{\code{UserService$findUserByEmail()}}
\item \href{#method-getSamlMessage}{\code{UserService$getSamlMessage()}}
\item \href{#method-cookieConsent}{\code{UserService$cookieConsent()}}
\item \href{#method-connect}{\code{UserService$connect()}}
\item \href{#method-connect2}{\code{UserService$connect2()}}
\item \href{#method-createUser}{\code{UserService$createUser()}}
\item \href{#method-hasUserName}{\code{UserService$hasUserName()}}
\item \href{#method-updatePassword}{\code{UserService$updatePassword()}}
\item \href{#method-updateBillingInfo}{\code{UserService$updateBillingInfo()}}
\item \href{#method-viesInfo}{\code{UserService$viesInfo()}}
\item \href{#method-summary}{\code{UserService$summary()}}
\item \href{#method-resourceSummary}{\code{UserService$resourceSummary()}}
\item \href{#method-profiles}{\code{UserService$profiles()}}
\item \href{#method-createToken}{\code{UserService$createToken()}}
\item \href{#method-isTokenValid}{\code{UserService$isTokenValid()}}
\item \href{#method-setTeamPrivilege}{\code{UserService$setTeamPrivilege()}}
\item \href{#method-getServerVersion}{\code{UserService$getServerVersion()}}
\item \href{#method-getInvited}{\code{UserService$getInvited()}}
\item \href{#method-sendValidationMail}{\code{UserService$sendValidationMail()}}
\item \href{#method-sendResetPasswordEmail}{\code{UserService$sendResetPasswordEmail()}}
\item \href{#method-changeUserPassword}{\code{UserService$changeUserPassword()}}
\item \href{#method-validateUser}{\code{UserService$validateUser()}}
\item \href{#method-canCreatePrivateProject}{\code{UserService$canCreatePrivateProject()}}
\item \href{#method-clone}{\code{UserService$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="create">}\href{../../tercenApi/html/HttpClientService.html#method-create}{\code{tercenApi::HttpClientService$create()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="delete">}\href{../../tercenApi/html/HttpClientService.html#method-delete}{\code{tercenApi::HttpClientService$delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="findKeys">}\href{../../tercenApi/html/HttpClientService.html#method-findKeys}{\code{tercenApi::HttpClientService$findKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="findStartKeys">}\href{../../tercenApi/html/HttpClientService.html#method-findStartKeys}{\code{tercenApi::HttpClientService$findStartKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="fromTson">}\href{../../tercenApi/html/HttpClientService.html#method-fromTson}{\code{tercenApi::HttpClientService$fromTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="get">}\href{../../tercenApi/html/HttpClientService.html#method-get}{\code{tercenApi::HttpClientService$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="getServiceUri">}\href{../../tercenApi/html/HttpClientService.html#method-getServiceUri}{\code{tercenApi::HttpClientService$getServiceUri()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="list">}\href{../../tercenApi/html/HttpClientService.html#method-list}{\code{tercenApi::HttpClientService$list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="onResponseError">}\href{../../tercenApi/html/HttpClientService.html#method-onResponseError}{\code{tercenApi::HttpClientService$onResponseError()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="toTson">}\href{../../tercenApi/html/HttpClientService.html#method-toTson}{\code{tercenApi::HttpClientService$toTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="update">}\href{../../tercenApi/html/HttpClientService.html#method-update}{\code{tercenApi::HttpClientService$update()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$new(baseRestUri, client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findTeamMembers"></a>}}
\if{latex}{\out{\hypertarget{method-findTeamMembers}{}}}
\subsection{Method \code{findTeamMembers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$findTeamMembers(keys = NULL, useFactory = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findUserByCreatedDateAndName"></a>}}
\if{latex}{\out{\hypertarget{method-findUserByCreatedDateAndName}{}}}
\subsection{Method \code{findUserByCreatedDateAndName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$findUserByCreatedDateAndName(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findUserByEmail"></a>}}
\if{latex}{\out{\hypertarget{method-findUserByEmail}{}}}
\subsection{Method \code{findUserByEmail()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$findUserByEmail(keys = NULL, useFactory = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSamlMessage"></a>}}
\if{latex}{\out{\hypertarget{method-getSamlMessage}{}}}
\subsection{Method \code{getSamlMessage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$getSamlMessage(type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cookieConsent"></a>}}
\if{latex}{\out{\hypertarget{method-cookieConsent}{}}}
\subsection{Method \code{cookieConsent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$cookieConsent(dummy)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect"></a>}}
\if{latex}{\out{\hypertarget{method-connect}{}}}
\subsection{Method \code{connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$connect(usernameOrEmail, password)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect2"></a>}}
\if{latex}{\out{\hypertarget{method-connect2}{}}}
\subsection{Method \code{connect2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$connect2(domain, usernameOrEmail, password)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createUser"></a>}}
\if{latex}{\out{\hypertarget{method-createUser}{}}}
\subsection{Method \code{createUser()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$createUser(user, password)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasUserName"></a>}}
\if{latex}{\out{\hypertarget{method-hasUserName}{}}}
\subsection{Method \code{hasUserName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$hasUserName(username)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updatePassword"></a>}}
\if{latex}{\out{\hypertarget{method-updatePassword}{}}}
\subsection{Method \code{updatePassword()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$updatePassword(userId, password)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateBillingInfo"></a>}}
\if{latex}{\out{\hypertarget{method-updateBillingInfo}{}}}
\subsection{Method \code{updateBillingInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$updateBillingInfo(userId, billingInfo)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-viesInfo"></a>}}
\if{latex}{\out{\hypertarget{method-viesInfo}{}}}
\subsection{Method \code{viesInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$viesInfo(country_code, vatNumber)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$summary(userId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resourceSummary"></a>}}
\if{latex}{\out{\hypertarget{method-resourceSummary}{}}}
\subsection{Method \code{resourceSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$resourceSummary(userId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-profiles"></a>}}
\if{latex}{\out{\hypertarget{method-profiles}{}}}
\subsection{Method \code{profiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$profiles(userId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createToken"></a>}}
\if{latex}{\out{\hypertarget{method-createToken}{}}}
\subsection{Method \code{createToken()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$createToken(userId, validityInSeconds)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isTokenValid"></a>}}
\if{latex}{\out{\hypertarget{method-isTokenValid}{}}}
\subsection{Method \code{isTokenValid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$isTokenValid(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTeamPrivilege"></a>}}
\if{latex}{\out{\hypertarget{method-setTeamPrivilege}{}}}
\subsection{Method \code{setTeamPrivilege()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$setTeamPrivilege(username, principal, privilege)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getServerVersion"></a>}}
\if{latex}{\out{\hypertarget{method-getServerVersion}{}}}
\subsection{Method \code{getServerVersion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$getServerVersion(module)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInvited"></a>}}
\if{latex}{\out{\hypertarget{method-getInvited}{}}}
\subsection{Method \code{getInvited()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$getInvited(email)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendValidationMail"></a>}}
\if{latex}{\out{\hypertarget{method-sendValidationMail}{}}}
\subsection{Method \code{sendValidationMail()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$sendValidationMail(email)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendResetPasswordEmail"></a>}}
\if{latex}{\out{\hypertarget{method-sendResetPasswordEmail}{}}}
\subsection{Method \code{sendResetPasswordEmail()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$sendResetPasswordEmail(email)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-changeUserPassword"></a>}}
\if{latex}{\out{\hypertarget{method-changeUserPassword}{}}}
\subsection{Method \code{changeUserPassword()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$changeUserPassword(token, password)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validateUser"></a>}}
\if{latex}{\out{\hypertarget{method-validateUser}{}}}
\subsection{Method \code{validateUser()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$validateUser(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-canCreatePrivateProject"></a>}}
\if{latex}{\out{\hypertarget{method-canCreatePrivateProject}{}}}
\subsection{Method \code{canCreatePrivateProject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$canCreatePrivateProject(teamOrUserId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
