% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_endpoint_service.R
\name{WorkerEndpointService}
\alias{WorkerEndpointService}
\title{WorkerEndpointService}
\format{
\code{\link{R6Class}} object.
}
\description{
WorkerEndpointService

WorkerEndpointService
}
\section{Methods}{

\describe{
}
}

\section{Super class}{
\code{\link[tercen:HttpClientService]{tercen::HttpClientService}} -> \code{WorkerEndpointService}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WorkerEndpointService$new()}}
\item \href{#method-findWorkerEndpointByOwner}{\code{WorkerEndpointService$findWorkerEndpointByOwner()}}
\item \href{#method-clone}{\code{WorkerEndpointService$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="create">}\href{../../tercen/html/HttpClientService.html#method-create}{\code{tercen::HttpClientService$create()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="delete">}\href{../../tercen/html/HttpClientService.html#method-delete}{\code{tercen::HttpClientService$delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="findKeys">}\href{../../tercen/html/HttpClientService.html#method-findKeys}{\code{tercen::HttpClientService$findKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="findStartKeys">}\href{../../tercen/html/HttpClientService.html#method-findStartKeys}{\code{tercen::HttpClientService$findStartKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="fromTson">}\href{../../tercen/html/HttpClientService.html#method-fromTson}{\code{tercen::HttpClientService$fromTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="get">}\href{../../tercen/html/HttpClientService.html#method-get}{\code{tercen::HttpClientService$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="getServiceUri">}\href{../../tercen/html/HttpClientService.html#method-getServiceUri}{\code{tercen::HttpClientService$getServiceUri()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="list">}\href{../../tercen/html/HttpClientService.html#method-list}{\code{tercen::HttpClientService$list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="onResponseError">}\href{../../tercen/html/HttpClientService.html#method-onResponseError}{\code{tercen::HttpClientService$onResponseError()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="toTson">}\href{../../tercen/html/HttpClientService.html#method-toTson}{\code{tercen::HttpClientService$toTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="update">}\href{../../tercen/html/HttpClientService.html#method-update}{\code{tercen::HttpClientService$update()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerEndpointService$new(baseRestUri, client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findWorkerEndpointByOwner"></a>}}
\if{latex}{\out{\hypertarget{method-findWorkerEndpointByOwner}{}}}
\subsection{Method \code{findWorkerEndpointByOwner()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerEndpointService$findWorkerEndpointByOwner(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkerEndpointService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
