% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_document_service.R
\name{ProjectDocumentService}
\alias{ProjectDocumentService}
\title{ProjectDocumentService}
\format{
\code{\link{R6Class}} object.
}
\description{
ProjectDocumentService

ProjectDocumentService
}
\section{Methods}{

\describe{
   \item{\code{getParentFolders(documentId)}}{method}
   \item{\code{cloneProjectDocument(documentId,projectId)}}{method}
}
}

\section{Super class}{
\code{\link[tercenApi:HttpClientService]{tercenApi::HttpClientService}} -> \code{ProjectDocumentService}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectDocumentService$new()}}
\item \href{#method-findProjectObjectsByLastModifiedDate}{\code{ProjectDocumentService$findProjectObjectsByLastModifiedDate()}}
\item \href{#method-findProjectObjectsByFolderAndName}{\code{ProjectDocumentService$findProjectObjectsByFolderAndName()}}
\item \href{#method-findFileByLastModifiedDate}{\code{ProjectDocumentService$findFileByLastModifiedDate()}}
\item \href{#method-findSchemaByLastModifiedDate}{\code{ProjectDocumentService$findSchemaByLastModifiedDate()}}
\item \href{#method-findSchemaByOwnerAndLastModifiedDate}{\code{ProjectDocumentService$findSchemaByOwnerAndLastModifiedDate()}}
\item \href{#method-findFileByOwnerAndLastModifiedDate}{\code{ProjectDocumentService$findFileByOwnerAndLastModifiedDate()}}
\item \href{#method-getParentFolders}{\code{ProjectDocumentService$getParentFolders()}}
\item \href{#method-cloneProjectDocument}{\code{ProjectDocumentService$cloneProjectDocument()}}
\item \href{#method-clone}{\code{ProjectDocumentService$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="create">}\href{../../tercenApi/html/HttpClientService.html#method-create}{\code{tercenApi::HttpClientService$create()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="delete">}\href{../../tercenApi/html/HttpClientService.html#method-delete}{\code{tercenApi::HttpClientService$delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="findKeys">}\href{../../tercenApi/html/HttpClientService.html#method-findKeys}{\code{tercenApi::HttpClientService$findKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="findStartKeys">}\href{../../tercenApi/html/HttpClientService.html#method-findStartKeys}{\code{tercenApi::HttpClientService$findStartKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="fromTson">}\href{../../tercenApi/html/HttpClientService.html#method-fromTson}{\code{tercenApi::HttpClientService$fromTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="get">}\href{../../tercenApi/html/HttpClientService.html#method-get}{\code{tercenApi::HttpClientService$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="getServiceUri">}\href{../../tercenApi/html/HttpClientService.html#method-getServiceUri}{\code{tercenApi::HttpClientService$getServiceUri()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="list">}\href{../../tercenApi/html/HttpClientService.html#method-list}{\code{tercenApi::HttpClientService$list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="onResponseError">}\href{../../tercenApi/html/HttpClientService.html#method-onResponseError}{\code{tercenApi::HttpClientService$onResponseError()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="toTson">}\href{../../tercenApi/html/HttpClientService.html#method-toTson}{\code{tercenApi::HttpClientService$toTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercenApi" data-topic="HttpClientService" data-id="update">}\href{../../tercenApi/html/HttpClientService.html#method-update}{\code{tercenApi::HttpClientService$update()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDocumentService$new(baseRestUri, client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findProjectObjectsByLastModifiedDate"></a>}}
\if{latex}{\out{\hypertarget{method-findProjectObjectsByLastModifiedDate}{}}}
\subsection{Method \code{findProjectObjectsByLastModifiedDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDocumentService$findProjectObjectsByLastModifiedDate(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findProjectObjectsByFolderAndName"></a>}}
\if{latex}{\out{\hypertarget{method-findProjectObjectsByFolderAndName}{}}}
\subsection{Method \code{findProjectObjectsByFolderAndName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDocumentService$findProjectObjectsByFolderAndName(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findFileByLastModifiedDate"></a>}}
\if{latex}{\out{\hypertarget{method-findFileByLastModifiedDate}{}}}
\subsection{Method \code{findFileByLastModifiedDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDocumentService$findFileByLastModifiedDate(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findSchemaByLastModifiedDate"></a>}}
\if{latex}{\out{\hypertarget{method-findSchemaByLastModifiedDate}{}}}
\subsection{Method \code{findSchemaByLastModifiedDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDocumentService$findSchemaByLastModifiedDate(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findSchemaByOwnerAndLastModifiedDate"></a>}}
\if{latex}{\out{\hypertarget{method-findSchemaByOwnerAndLastModifiedDate}{}}}
\subsection{Method \code{findSchemaByOwnerAndLastModifiedDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDocumentService$findSchemaByOwnerAndLastModifiedDate(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findFileByOwnerAndLastModifiedDate"></a>}}
\if{latex}{\out{\hypertarget{method-findFileByOwnerAndLastModifiedDate}{}}}
\subsection{Method \code{findFileByOwnerAndLastModifiedDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDocumentService$findFileByOwnerAndLastModifiedDate(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getParentFolders"></a>}}
\if{latex}{\out{\hypertarget{method-getParentFolders}{}}}
\subsection{Method \code{getParentFolders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDocumentService$getParentFolders(documentId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cloneProjectDocument"></a>}}
\if{latex}{\out{\hypertarget{method-cloneProjectDocument}{}}}
\subsection{Method \code{cloneProjectDocument()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDocumentService$cloneProjectDocument(documentId, projectId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDocumentService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
